requirements.txt
python-chess==1.999
pygame==2.5.2